# Conflicts:
#	platform/core-ui/src/util/IconUtil.kt

+++ platform/platform-impl/src/com/intellij/diagnostic/dropErrorActions.kt
@@ -15,9 +15,8 @@ import com.intellij.openapi.util.text.StringUtil
 import com.intellij.util.TimeoutUtil
 import java.awt.event.ActionEvent.CTRL_MASK
 import java.awt.event.ActionEvent.SHIFT_MASK
-import java.io.*
+import java.io.RandomAccessFile
 import java.util.*
-
 
 private const val TEST_LOGGER = "TEST.LOGGER"
 private const val TEST_MESSAGE = "test exception; please ignore"
@@ -66,9 +65,7 @@ internal class DropAnErrorWithAttachmentsAction : DumbAwareAction("Drop an Error
     val size = 300 * 1024 * 1024
     val file = FileUtil.createTempFile("large-attachment", ".bin", true)
     RandomAccessFile(file, "rw").apply { setLength(size.toLong()) }
-
-    return arrayOf(Attachment("large.txt", file,
-                              "A large attachment of size: $size bytes").apply { isIncluded = true })
+    return arrayOf(Attachment("large.txt", file, "A large attachment of size: $size bytes").apply { isIncluded = true })
   }
 }
 
@@ -124,6 +121,6 @@ internal class SimulateFreeze : DumbAwareAction("Simulate a Freeze") {
 
   // Keep it a function to detect it in EA
   private fun simulatedFreeze(ms: Long) {
-    Thread.sleep(ms.toLong())
+    Thread.sleep(ms)
   }
 }
